+++ model/schema_combo.go
@@ -49,7 +49,7 @@ func (s *Schema[T]) getOrCreate(rq *my.Request, p *GetOrCreateParams[T], isTx bo
 	numRows := len(rows)
 	if numRows > 1 {
 		rq.Fail(my.Err404, "Failed to get one %s", p.Name)
-		err := fmt.Errorf("public: Multiple %s found", p.Name)
+		err = fmt.Errorf("public: Multiple %s found", p.Name)
 		if isTx {
 			err = qb.Rollback(rq.Tx, err) // manual rollback
 		}