+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.3.26 - Clock Functions
+  * Commit: 2026-04-28 14:23
+  * Clock Extract Date Parts
+  * Time Now Functions
+  * Time Parsing and Expiry Functions
 ## v0.3.25 - Fork Name 
   * Commit: 2026-04-24 19:50
   * ForkName in ForkData and ForkAction